pyproject.toml
[project]
name = "ai-services"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dspy>=3.0.4",
    "litellm>=1.80.5",
    "numpy>=2.3.5",
    "sentence-transformers>=5.1.2",
    "weaviate-client>=4.18.1",
]
